A 64-bit buffer is faster on x86 for decoding It hurts encoding performance a bit, but improves decoder performance more so. For enwik8, performance with a 32- and 64-bit buffer on x86:
32-bit encoding: 1.560985 seconds
32-bit decoding: 2.429251 seconds

64-bit encoding: 1.708929 seconds
64-bit decoding: 1.807045 seconds

+++ fly/coders/bit_stream.h
@@ -11,7 +11,7 @@ namespace fly {
 
 typedef std::uint8_t byte_type;
 typedef std::uint16_t word_type;
-typedef std::size_t buffer_type;
+typedef std::uint64_t buffer_type;
 
 /**
  * Base class for writing to and reading from a binary stream. The first byte of
